welcome

+++ Service/TogetherService/PaintData.cs
@@ -33,8 +33,8 @@ namespace TogetherService
 
     public class PaintCoordinate
     {
-        public int OffsetX { get; set; }
+        public float OffsetX { get; set; }
 
-        public int OffsetY { get; set; }
+        public float OffsetY { get; set; }
     }
 }
